feat: [IXSPD1-2000] support airdrop by csv file without header
