+++ src/components/Uploader/index.tsx
@@ -2,7 +2,6 @@ import React, { ChangeEvent, FC, useState } from 'react'
 import { Box, Flex } from 'rebass'
 import { FileWithPath } from 'react-dropzone'
 import { TYPE } from 'theme'
-import Big from 'big.js'
 import { Label } from 'components/Label'
 import Upload from 'components/Upload'
 import { FilePreview, FilePreviewPayout } from 'components/FilePreview'
@@ -12,12 +11,6 @@ import { ReactComponent as UploadLogo } from 'assets/images/NewDownloads.svg'
 import { Text } from 'rebass'
 import styled, { css } from 'styled-components'
 import { UploaderCard } from 'pages/KYC/styleds'
-import { AirdropTable, Body, Header } from 'pages/CreateAirdropPayoutEvent/AirdropTable'
-import { Pagination } from 'components/Pagination'
-import { adminOffset } from 'state/admin/constants'
-import Papa from 'papaparse'
-import { ethers } from 'ethers'
-import { useShowError } from 'state/application/hooks'
 
 export interface UploaderProps {
   files: FileWithPath[]
@@ -30,14 +23,14 @@ export interface UploaderProps {
   required?: boolean
   tooltipText?: string | JSX.Element
   isDisabled?: boolean
+  isUploadHide?: boolean
   id?: any
   name?: string
   onChange?: (e: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => void
   isPayoutpage?: boolean
   acceptedFileTypes?: string[]
-  setTotalWallets?: (value: number) => void
-  setTotalAmount?: (value: string) => void
   onCsvRowsChange?: (value: string[][]) => void
+  onCsvFileUpload?: (text: string) => void
 }
 
 export const Uploader: FC<UploaderProps> = ({
@@ -52,84 +45,23 @@ export const Uploader: FC<UploaderProps> = ({
   optional = false,
   tooltipText,
   isDisabled = false,
+  isUploadHide = false,
   isPayoutpage,
   acceptedFileTypes,
-  setTotalWallets,
-  setTotalAmount,
   onCsvRowsChange,
+  onCsvFileUpload,
 }: UploaderProps) => {
-  const [csvRows, setCsvRows] = useState<string[][]>([])
   const defaultAcceptedFiles = `${AcceptFiles.PDF},image/jpeg,image/png`
   const accept = acceptedFileTypes?.join(',') || defaultAcceptedFiles
-  const [currentPage, setCurrentPage] = useState(1)
-  const showError = useShowError()
-  // const [hasError, setHasError] = useState(false)
 
   const handleFileUpload = (file: FileWithPath) => {
     if (file.type === 'text/csv') {
       const reader = new FileReader()
       reader.onload = (event) => {
         const text = event.target?.result as string
-        uploadCSVFile(text)
+        onCsvFileUpload?.(text)
       }
       reader.readAsText(file)
-    }
-  }
-
-  const uploadCSVFile = (csvText: string) => {
-    const parsed = Papa.parse<string[]>(csvText, { delimiter: ',', skipEmptyLines: true }).data
-    let hasError = false
-
-    if (parsed.length === 0) {
-      showError('Invalid file: The file is empty.')
-      return
-    }
-
-    const [headers, ...rows] = parsed
-
-    if (headers[0].trim() !== 'walletAddress' || headers[1].trim() !== 'amount') {
-      showError('Invalid file: The file headers must be "walletAddress" and "amount"')
-      setCsvRows([])
-      onCsvRowsChange?.([])
-      setTotalWallets?.(0)
-      setTotalAmount?.('0')
-      return
-    }
-
-    const validatedRows = rows
-      .map(([address, amount]) => {
-        const trimmedAddress = address?.trim()
-        const trimmedAmount = amount?.trim()?.replace(/,/g, '')
-
-        if (!ethers.utils.isAddress(trimmedAddress)) {
-          console.error(`Invalid address: ${trimmedAddress}`)
-          hasError = true
-          return null
-        }
-
-        if (isNaN(Number(trimmedAmount))) {
-          console.error(`Invalid amount: ${trimmedAmount}`)
-          hasError = true
-          return null
-        }
-
-        return [trimmedAddress, trimmedAmount]
-      })
-      .filter((row) => row !== null) as string[][]
-
-    if (hasError) {
-      showError('Invalid file: please ensure all addresses are valid and amounts are correctly formatted.')
-      setCsvRows([])
-      onCsvRowsChange?.([])
-      setTotalWallets?.(0)
-      setTotalAmount?.('0')
-    } else {
-      setCsvRows(validatedRows)
-      onCsvRowsChange?.(validatedRows)
-      const totalWallets = validatedRows.length
-      const totalAmount = validatedRows.reduce((sum, [, amount]) => sum.add(new Big(amount)), new Big(0))
-      setTotalWallets?.(totalWallets)
-      setTotalAmount?.(totalAmount.toString())
     }
   }
 
@@ -144,18 +76,7 @@ export const Uploader: FC<UploaderProps> = ({
 
   const handleFileDelete = (index: number) => {
     handleDeleteClick(index)
-    setCsvRows([])
     onCsvRowsChange?.([])
-  }
-
-  const onPageChange = (page: number) => {
-    event?.preventDefault()
-    const totalPages = Math.ceil(csvRows.length / adminOffset)
-
-    if (page > 0 && page <= totalPages) {
-      setCurrentPage(page)
-      window.scrollTo({ top: 0, behavior: 'smooth' })
-    }
   }
 
   return (
@@ -203,7 +124,7 @@ export const Uploader: FC<UploaderProps> = ({
         </Flex>
       )}
 
-      {!isDisabled && csvRows.length === 0 && (
+      {!isDisabled && !isUploadHide && (
         <Upload isDisabled={isDisabled} accept={accept as AcceptFiles} data-testid={id} file={null} onDrop={handleDrop}>
           <UploaderCard>
             <Flex flexDirection="column" justifyContent="center" alignItems="center" style={{ maxWidth: 100 }}>
@@ -217,21 +138,6 @@ export const Uploader: FC<UploaderProps> = ({
             </Flex>
           </UploaderCard>
         </Upload>
-      )}
-
-      {csvRows.length > 0 && (
-        <>
-          <AirdropTable
-            header={<Header columns={['Wallet Address', 'Amount']} />}
-            body={<Body rows={csvRows.slice((currentPage - 1) * adminOffset, currentPage * adminOffset)} />}
-          />
-          <Pagination
-            totalPages={Math.ceil(csvRows.length / adminOffset)}
-            page={currentPage}
-            onPageChange={onPageChange}
-            totalItems={csvRows.length}
-          />
-        </>
       )}
 
       {error && (
